Implemented droping items from inventory
also observed a bug where I was put back on the first floor
from the second floor. When I returned to the second floor
some monsters couldn't track me down. But not the case with
all monsters on the second floor.

When I dead the game crashes, becourse of a overflow in the
entityManager.