Implemented wandering for Entities. Implemented an overload for IComponent_Entity with allow for wandering
Currently only the rat can wander, and it only starts when in FOV
of the player. Should be all entities.

the _goalMap has 3 goals in the beginning, should be refilled when
the entity reaches the goal.

Implemented a getPlayerDistance method in Entity.cs it calculates
actual distance, should be tile distance.

need to improve mock-up gui to better understand what is going on.

Currently the wandering monsters ignore other monsters, think i can
make the tile unwalkable and seethrough.

don't know if every entity should have a goalmap, or if it should be
a static.