Added a dedicated StatusScreen Class
This class will control all gui elements related to player status, and will contain a overview of all visable objects with in the player fov.
Hovering over an entry will show the path as if the mouse was hovering over the entry.
Clicking the entry will make the player move towards the entry as if clicked on it.